pyproject.toml
[project]
name = "hkex-settlement-parser"
version = "0.1.0"
description = "HKEX Settlement Price File Parser with FastAPI backend"
authors = [
    {name = "HKEX Settlement Team", email = "team@hkex.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "redis>=5.0.0",
    "influxdb-client>=1.38.0",
    "cassandra-driver>=3.28.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "behave>=1.2.7",
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
